Update 01_introduction.asciidoc
Fixed some typos and small errors mostly.
The capitalization of heading was not consistent.  I've gone with capitalizing the first word and not capitalizing the remaining words in the heading as this was the vast majority of the headings.

So: Lightning Network use cases, users, and their stories
Not: Lightning Network Use Cases, Users, and Their Stories

I'm going through the rest of the chapters and I'll try to make them consistent.

One questions - is this Chapter 1 or the Introduction?  If it is an introduction, it should be "In this introduction" and not "In this chapter".